WIP web: Proper header at MdRaidEditor
